pyproject.toml
[project]
name = "data-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "polars>=1.19.0",
    "python-binance>=1.0.27",
    "requests>=2.32.3",
    "scipy>=1.15.1",
]
